FileListクラスを削除して、Displayクラスにメソッドを持たせる
