+++ 08.ls_object/lib/Ls/display.rb
@@ -4,13 +4,17 @@ require_relative 'file_list'
 
 module Ls
   class Display
+    COLUMN_NUMBER = 3
+
     def initialize(options)
-      @file_list = Ls::FileList.new(options)
+      file_names = options['a'] ? Dir.glob('*', ::File::FNM_DOTMATCH) : Dir.glob('*')
+      file_names = file_names.reverse if options['r']
+      @file_list = file_names.map { |file_name| Ls::File.new(file_name) }
     end
 
     def show_list_short
-      devided_list = @file_list.devide_the_file_list
-      max_filename_length = @file_list.calc_max_file_name_length
+      devided_list = devide_the_file_list
+      max_filename_length = calc_max_file_name_length
       # 表示の間隔は常に8の倍数
       adjustment_width = (max_filename_length / 8 + 1) * 8
 
@@ -21,6 +25,19 @@ module Ls
         end.join
         puts line
       end
+    end
+
+    private
+
+    def devide_the_file_list
+      num_to_slice = (@file_list.size.to_f / COLUMN_NUMBER).ceil
+      devided_list = @file_list.each_slice(num_to_slice).to_a
+      devided_list.last << nil while devided_list.last.size < num_to_slice
+      devided_list
+    end
+
+    def calc_max_file_name_length
+      @file_list.max_by(&:name_length).name_length
     end
   end
 end
